Test implementation of approach for supporting assurance claim primitives like `targets_reported`
Relies on the scene.dynamicScenario._dummyNamespace component of the
scene argument to the Simulator constructor. The former provides full
access to all scenario file variables, allowing them to be updated
from the simulator. The require statements and monitors allow scenario
file variables to be used freely, even in expressions involving list
comprehensions.